feature_extracion/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
appdirs=1.4.4=pypi_0
async-timeout=4.0.1=pypi_0
attrs=21.2.0=pyhd3eb1b0_0
audiomentations=0.20.0=pypi_0
audioread=2.1.9=pypi_0
autobahn=21.11.1=pyhd8ed1ab_0
autocfg=0.0.8=pypi_0
automat=20.2.0=py_0
bcrypt=3.2.0=py39he8ac12f_0
blas=1.0=mkl
blosc=1.21.0=h9c3ff4c_0
bottleneck=1.3.2=py39hdd57654_1
brotli=1.0.9=he6710b0_2
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
c-blosc2=2.0.4=h5f21a17_1
ca-certificates=2022.3.29=h06a4308_0
cairo=1.16.0=hf32fb01_1
certifi=2021.10.8=py39h06a4308_2
cffi=1.15.0=py39h4bc2ebd_0
cfitsio=4.0.0=h9a35b8e_0
charls=2.2.0=h2531618_0
charset-normalizer=2.0.7=pypi_0
click=8.0.3=pypi_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyh9f0ad1d_0
configparser=5.1.0=pypi_0
constantly=15.1.0=pyh2b92418_0
cryptography=35.0.0=py39hd23ed53_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.0=py39h27cfd23_0
dask-core=2021.10.0=pyhd3eb1b0_0
data=0.4=pypi_0
datasets=1.16.1=pypi_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=pypi_0
dill=0.3.4=pypi_0
dlib=19.22.0=py39hcec8a4c_0
docker-pycreds=0.4.0=pypi_0
einops=0.3.2=pypi_0
expat=2.4.1=h2531618_2
ffmpeg=4.3.2=hca11adc_1
filelock=3.4.0=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.28.1=pypi_0
freeglut=3.0.0=hf484d3e_5
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=pypi_0
fsspec=2021.11.1=pypi_0
funcsigs=1.0.2=pypi_0
future=0.18.2=pypi_0
gettext=0.21.0=hf68c758_0
gevent=21.8.0=pypi_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pypi_0
gitpython=3.1.24=pypi_0
glib=2.70.1=h780b84a_0
glib-tools=2.70.1=h780b84a_0
gluoncv=0.10.4.post4=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
graphite2=1.3.14=h23475e2_0
greenlet=1.1.2=pypi_0
gst-plugins-base=1.18.5=hc549b11_0
gstreamer=1.18.5=ha1a6a79_0
harfbuzz=3.1.1=h83ec7ef_0
hdf5=1.12.1=nompi_h2750804_103
huggingface-hub=0.1.2=pypi_0
hyperlink=21.0.0=pyhd3eb1b0_0
icu=68.2=h9c3ff4c_0
idna=3.3=pyhd3eb1b0_0
imagecodecs=2021.11.20=py39h571908b_1
imagehash=4.2.1=pyhd3eb1b0_0
imageio=2.11.0=pypi_0
imageio-ffmpeg=0.4.5=pypi_0
imgaug=0.4.0=pypi_0
incremental=21.3.0=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
jasper=2.0.14=hd8c5072_2
jbig=2.1=hdba287a_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jxrlib=1.1=h7b6447c_2
kiwisolver=1.3.2=pypi_0
krb5=1.19.2=hcc1bbae_3
lame=3.100=h7b6447c_0
latex=0.7.0=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h295c915_0
libaec=1.0.6=h9c3ff4c_0
libblas=3.9.0=12_linux64_mkl
libbrotlicommon=1.0.9=h7f98852_6
libbrotlidec=1.0.9=h7f98852_6
libbrotlienc=1.0.9=h7f98852_6
libcblas=3.9.0=12_linux64_mkl
libclang=11.1.0=default_ha53f305_1
libcurl=7.80.0=h2574ce0_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h7f8727e_1
libevent=2.1.10=h9b69904_4
libffi=3.4.2=h295c915_2
libgcc-ng=11.2.0=h1d223b6_11
libgfortran-ng=11.2.0=h69a702a_11
libgfortran5=11.2.0=h5c6108e_11
libglib=2.70.1=h174f98d_0
libglu=9.0.0=hf484d3e_1
libiconv=1.16=h516909a_0
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=12_linux64_mkl
liblapacke=3.9.0=12_linux64_mkl
libllvm11=11.1.0=h3826bc1_0
libnghttp2=1.46.0=hce63b2e_0
libogg=1.3.5=h27cfd23_1
libopencv=4.5.3=py39ha7f30a5_5
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libpq=13.3=hd57d9b9_3
libprotobuf=3.18.1=h780b84a_0
librosa=0.8.1=pypi_0
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=he4da1e4_11
libtasn1=4.16.0=h27cfd23_0
libtiff=4.3.0=h6f004c6_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvorbis=1.3.7=h7b6447c_0
libwebp-base=1.2.1=h7f98852_0
libxcb=1.14=h7b6447c_0
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
libxslt=1.1.33=h15afd5d_2
libzlib=1.2.11=h36c2ea0_1013
libzopfli=1.0.3=he6710b0_0
llvm-openmp=12.0.1=h4bd325d_1
llvmlite=0.37.0=pypi_0
locket=0.2.1=py39h06a4308_1
lxml=4.8.0=py39hb9d737c_1
lz4-c=1.9.3=h295c915_1
matplotlib-base=3.5.0=py39h3ed280b_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
moviepy=1.0.3=pypi_0
multidict=5.2.0=pypi_0
multiprocess=0.70.12.2=pypi_0
munkres=1.1.4=py_0
mxnet-cu100mkl=1.5.1.post0=pypi_0
mxnet-cu92=1.7.0=pypi_0
mysql-common=8.0.27=ha770c72_1
mysql-libs=8.0.27=hfa10184_1
ncurses=6.2=h58526e2_4
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pyhd3eb1b0_0
nltk=3.6.7=pyhd8ed1ab_0
nspr=4.32=h9c3ff4c_1
nss=3.72=hb5efdd6_0
numba=0.54.1=pypi_0
numexpr=2.7.3=py39h22e1b3c_1
numpy=1.20.3=pypi_0
numpy-base=1.21.2=py39h79a1101_0
olefile=0.46=pyhd3eb1b0_0
opencv=4.5.3=py39hf3d152e_5
opencv-python=4.5.4.60=pypi_0
openface=0.2.1=pypi_0
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1n=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.4=py39h8c16a72_0
partd=1.2.0=pyhd3eb1b0_0
pathtools=0.1.2=pypi_0
pcre=8.45=h295c915_0
pillow=8.4.0=py39ha612740_0
pip=21.2.4=py39h06a4308_0
pixman=0.40.0=h7f8727e_1
pooch=1.5.2=pypi_0
portalocker=2.3.2=pypi_0
proglog=0.1.9=pypi_0
promise=2.3=pypi_0
protobuf=3.19.1=pypi_0
psutil=5.8.0=pypi_0
py-opencv=4.5.3=py39hef51801_5
pyarrow=6.0.1=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pyhamcrest=2.0.2=pyhd3eb1b0_2
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.6=pypi_0
python=3.9.7=hb7a2778_3_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-graphviz=0.8.4=pypi_0
python_abi=3.9=2_cp39
pytorch=1.10.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.2.0=pypi_0
pyyaml=6.0=pypi_0
qt=5.12.9=hda022c4_4
readline=8.1=h27cfd23_0
regex=2021.11.10=pypi_0
requests=2.26.0=pypi_0
resampy=0.2.2=pypi_0
sacremoses=0.0.46=pypi_0
scikit-image=0.18.3=pypi_0
scikit-learn=1.0.1=py39h51133e4_0
scipy=1.7.2=pypi_0
seaborn=0.11.2=pypi_0
sentry-sdk=1.5.0=pypi_0
service_identity=18.1.0=pyhd3eb1b0_1
setproctitle=1.2.2=pypi_0
setuptools=58.0.4=py39h06a4308_0
setuptools-scm=6.3.2=pypi_0
shapely=1.8.0=pypi_0
shortuuid=1.0.8=pypi_0
shutilwhich=1.1.0=pypi_0
six=1.16.0=pyhd3eb1b0_0
smmap=5.0.0=pypi_0
snappy=1.1.8=he6710b0_0
soundfile=0.10.3.post1=pypi_0
specaugment=1.2.5=pypi_0
sqlite=3.36.0=hc218d9a_0
subprocess32=3.5.4=pypi_0
tempdir=0.7.1=pypi_0
termcolor=1.1.0=pypi_0
textblob=0.15.3=py_0
threadpoolctl=3.0.0=pypi_0
tifffile=2021.11.2=pypi_0
timm=0.4.12=pypi_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.10.3=pypi_0
tomli=1.2.2=pypi_0
toolz=0.11.2=pyhd3eb1b0_0
torchaudio=0.10.0=py39_cu113
torchvision=0.11.1=pypi_0
tqdm=4.62.3=pypi_0
transformers=4.12.5=pypi_0
twisted=21.7.0=py39h7f8727e_0
txaio=21.2.1=pyhd8ed1ab_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
urllib3=1.26.7=pypi_0
wandb=0.12.11=pypi_0
websocket=0.2.1=pypi_0
websocket-client=1.2.3=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
wordcloud=1.8.1=pypi_0
x264=1!161.3030=h7f98852_1
xxhash=2.0.2=pypi_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
yaml=0.2.5=h7b6447c_0
yarl=1.7.2=pypi_0
yaspin=2.1.0=pypi_0
zfp=0.5.5=h2531618_6
zlib=1.2.11=h36c2ea0_1013
zope=1.0=py39h06a4308_1
zope-event=4.5.0=pypi_0
zope.interface=5.4.0=py39h7f8727e_0
zstd=1.5.0=ha4553b6_1
